refactor: Move layout methods to dedicated directory.
