+++ packages/vgplot/src/layout/index.js
@@ -1,0 +1,2 @@
+export { hconcat, vconcat } from './concat.js';
+export { hspace, vspace } from './space.js';
